Fix api source file selector
Using the package root directory as the base for a `find` tree results
in the entire root directory being initially selected. When broccoli's
server is then run, it was watching the entire package directory,
including tmp (which immediately changed thanks to the build).

Now, it selects only the lib folder directly, avoiding the problem.